Remove uint16_t and address-mask changes; keep only the three valid bug fixes
Agent-Logs-Url: https://github.com/dgatf/I2C-slave-multi-address-RP2040/sessions/7d7b83b8-f2c5-4f72-b92c-5298e23d7735

+++ arduino/i2c_multi/i2c_multi.ino
@@ -51,7 +51,7 @@ void i2c_request_handler(uint8_t address) {
     }
 }
 
-void i2c_stop_handler(uint16_t length) {
+void i2c_stop_handler(uint8_t length) {
     sprintf(str_out, "\nTotal bytes: %u", length);
     Serial.print(str_out);
 }
